Fixed homepage, tasks
